+++ build_cli/test/src/pubviz_example.g.dart
@@ -6,12 +6,17 @@ part of 'pubviz_example.dart';
 // CliGenerator
 // **************************************************************************
 
-T _$enumValueHelper<T>(String enumName, List<T> values, String enumValue) =>
-    enumValue == null
-        ? null
-        : values.singleWhere((e) => e.toString() == '$enumName.$enumValue',
-            orElse: () => throw StateError(
-                'Could not find the value `$enumValue` in enum `$enumName`.'));
+T _$enumValueHelper<T>(Map<T, String> enumValues, String source) {
+  if (source == null) {
+    return null;
+  }
+  return enumValues.entries
+      .singleWhere((e) => e.value == source,
+          orElse: () => throw ArgumentError(
+              '`$source` is not one of the supported values: '
+              '${enumValues.values.join(', ')}'))
+      .key;
+}
 
 T _$badNumberFormat<T extends num>(
         String source, String type, String argName) =>
@@ -19,8 +24,8 @@ T _$badNumberFormat<T extends num>(
         'Cannot parse "$source" into `$type` for option "$argName".');
 
 PubvizOptions _$parsePubvizOptionsResult(ArgResults result) => PubvizOptions(
-    format: _$enumValueHelper(
-        'FormatOptions', FormatOptions.values, result['format'] as String),
+    format:
+        _$enumValueHelper(_$FormatOptionsEnumMap, result['format'] as String),
     secret: result['secret'] as String,
     ignorePackages: result['ignore-packages'] as List<String>,
     productionPort: int.tryParse(result['production-port'] as String) ??
@@ -36,6 +41,11 @@ PubvizOptions _$parsePubvizOptionsResult(ArgResults result) => PubvizOptions(
   ..listOfNothing = result['list-of-nothing'] as List
   ..listOfDynamic = result['list-of-dynamic'] as List
   ..listOfObject = result['list-of-object'] as List;
+
+const _$FormatOptionsEnumMap = <FormatOptions, String>{
+  FormatOptions.dot: 'dot',
+  FormatOptions.html: 'html'
+};
 
 ArgParser _$populatePubvizOptionsParser(ArgParser parser) => parser
   ..addOption('format', abbr: 'f', defaultsTo: 'html', allowed: [